refactor: user 패키지 이름 member로 변경 [#1]
