+++ src/main/java/com/example/sns/member/infrastructure/MemberRepository.java
@@ -1,12 +1,12 @@
-package com.example.sns.user.infrastructure;
+package com.example.sns.member.infrastructure;
 
-import com.example.sns.user.domain.Member;
+import com.example.sns.member.domain.Member;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 import java.util.Optional;
 
 @Repository
-public interface UserRepository extends JpaRepository<Member, Long> {
+public interface MemberRepository extends JpaRepository<Member, Long> {
     Optional<Member> findBySocialId(String socialId);
 }
